Moved the claims component from the core directory to component, changed the default values of amount and currency in draft, Made the textContent of the transaction Date component dynamic
