+++ src/components/Claims/index.tsx
@@ -1,9 +1,6 @@
 import {  SafeAreaView, Text, View } from 'react-native';
 import React from 'react';
 
-
-
-
 const Claims = ({data,color,symbol}) => {
  return (
    <SafeAreaView>
